get rid of cast warning
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@5568 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/gist/analysiscursor.cc
@@ -100,8 +100,8 @@ namespace Gecode { namespace Gist {
   DistributeCursor::processCurrentNode(void) {
     VisualNode* n = node();
     if (n->getParent() != NULL) {
-      int h = (((double)n->getHeat() - (double)minHeat) / 
-               ((double)maxHeat-(double)minHeat)) * 179;
+      int h = (int)((((double)n->getHeat() - (double)minHeat) / 
+               ((double)maxHeat-(double)minHeat)) * 179);
       n->setHeat(h);
     } else {
       n->setHeat(0);
